pyproject.toml
[project]
name = "report-killer"
version = "0.2.0"
description = "An AI agent that reads Word documents and fills in answers to questions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.scripts]
report-killer = "report_killer.cli:main"

[tool.setuptools]
packages = ["report_killer"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
